+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.70] - 2024-03-03
+### Added
+* `LayerMaskExtensions` file for `LayerMask` convenience functions
+### Deprecated
+* `GameObjectExtensions.IsInLayerMask(LayerMask)` in favor of `LayerMaskExtensions.Contains(GameObject)`
+
 ## [0.0.69] - 2024-02-26
 ### Added
 * `GraphicsSettingsManager` toggle for the last graphics setting that was selected (cleared whenever there's a domain reload)